Dodano: - przykładową trasę "Mosty", - aktywność z mapą, - ustawianie markerów do atrakcji z trasy na mapie, - wroclaw.png przerzucone do drawable-xxhdpi - teraz działa na emulatorze i na telefonie.

+++ app/src/main/java/com/wroguide/model/PlaceDatabaseDAO.java
@@ -8,7 +8,7 @@ import java.util.Optional;
  */
 public class PlaceDatabaseDAO implements PlaceDAO {
     @Override
-    public Optional<Place> get(long id) {
+    public Optional<Place> get(int id) {
         return null;
     }
