+++ app/src/main/java/com/wroguide/model/PlaceFakeDAO.java
@@ -23,7 +23,7 @@ public class PlaceFakeDAO implements PlaceDAO {
     }
 
     @Override
-    public Optional<Place> get(long id) {
+    public Optional<Place> get(int id) {
         return null;
     }
